server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.9.2
motor==3.6.1
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
